Support downgrade to HTTP 1.1 for subrequests
When accessing a location that needs a client certificate, a downgrade
to HTTP 1.1 may be necessary if the current connection uses HTTP/2.
This must be done because HTTP/2 does not allow SSL renegotiations.

mod_h2 already handles this case, it sets a request note
"ssl-renegotiate-forbidden" and later evaluates it in
h2_headers_rcreate(). An RST_STREAM is sent to the client with the
error code HTTP_1_1_REQUIRED.

This did not work if there are custom error pages configured with
ErrorDocument. To show these error pages, Apache creates an internal
redirection in ap_die_r(). This creates a new internal request which
does not have the "ssl-renegotiate-forbidden" note.

Bugfix: Also look for the request note in linked requests.

Fixes #172